requirements.txt
langchain
llama-index
ollama
pytesseract
pyocr
onnxruntime
pyTelegramBotAPI
telebot
requests
Pillow
opencv-python
numpy
openai
llama-index-llms-ollama
llama-index-embeddings-huggingface
python-dotenv